Fix accuracy of geographic coordinates.

+++ MapBul.DBContext/marker.cs
@@ -48,9 +48,9 @@ namespace MapBul.DBContext
 
         public int BaseCategoryId { get; set; }
 
-        public float Lat { get; set; }
+        public double Lat { get; set; }
 
-        public float Lng { get; set; }
+        public double Lng { get; set; }
 
         [Required]
         [StringLength(200)]
